Fixing a type warning; NFC.

+++ src/test/Module.test.cpp
@@ -143,7 +143,7 @@ TEST(Unit_Module, protobufRoundTrip) {
   proto::Module Message;
 
   UUID ByteMapID, SymbolID, BlockID, DataID, SectionID;
-  int WhichSymbolic;
+  size_t WhichSymbolic;
 
   {
     Module *Original = Module::Create(Ctx);
